ai-service/requirements.txt
flask
flask-cors
rembg
pillow
